Changes from the review
